+++ src/Main.java
@@ -12,7 +12,7 @@ import javax.swing.JTextField;
 
 public class Main{
 
-    public static final int SIDE = 600 ;
+    public static final Double SIDE = 600.0 ;
     public static int no_players ;
     public static String side ;
     public static client[] sent_client ;
@@ -58,7 +58,7 @@ public class Main{
         //adding pong panel to the centre of the layout
         frame.add(pongPanel, BorderLayout.CENTER);
         //Setting the size to 600 as declared above
-        frame.setSize(SIDE, SIDE);
+        frame.setSize(SIDE.intValue(), SIDE.intValue());
         //setting the visibility to on
         frame.setVisible(true);
         //We wont be able to resize the frame
